refactor: tidy up SettingTableViewController
