+++ Imagine/Views/Cells/Community Header & Footers/AddOnCollectionViewFooter.swift
