+++ Imagine/Views/Cells/Community Header & Footers/AddOnHeaderView.swift
