+++ Imagine/Views/Cells/Community Header & Footers/TopicCollectionFooter.swift
